_cloudFunctions/requirements.txt
functions-framework==3.*
google-cloud
google-cloud-aiplatform
google-cloud-pubsub
google-cloud-bigquery
google-cloud-storage
google-cloud-logging
typing